backend/requirements.txt
# Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.1
aiofiles>=23.2.1
python-magic>=0.4.27
httpx>=0.25.2
tenacity>=8.2.3
structlog>=23.2.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0

# Integration SDKs (optional - install as needed)
stripe>=7.0.0
python-quickbooks>=0.9.4
intuitlib>=1.1.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.1
